backend/requirements.txt
alembic==1.14.1
apscheduler==3.10.4
aiosqlite==0.20.0
asyncpg==0.30.0
email-validator==2.2.0
fastapi==0.115.6
passlib[argon2,bcrypt]==1.7.4
psycopg[binary]==3.2.4
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
razorpay==1.4.2
redis==5.2.1
slowapi==0.1.9
sqlalchemy[asyncio]==2.0.36
uvicorn[standard]==0.34.0
groq==0.8.0
pandas==2.2.2
